feat(safetybell) : #55 - Add find within distance safetybell
- Add find within distance safetybel

Related to: #55

+++ backend/safe-service/src/main/java/com/safeservice/domain/safetybell/service/SafetyBellService.java
@@ -10,6 +10,6 @@ public interface SafetyBellService {
 
     void delete(SafetyBell safetyBell);
 
-    List<SafetyBell> findWithinDistance(double longitude, double latitude, int distance);
+    List<SafetyBell> findWithinDistance(double longitude, double latitude, double distance);
 
 }
